+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.21.1-1.6.7
+
+- Make it so recipes can specify a blockUnderneath, a block that will need to be anywhere under the Miner Controller to be able to get the items (no default recipes use it, it's mainly for modpack makers)
+  - KubeJS can also register those recipes, see the [wiki](https://github.com/Minimaxi28/VoidMiners-Remastered-Reforked/wiki#recipe-syntax)
+
 ## 1.21.1-1.6.6
 
 - Optimization of Miners and Solars